Customize Context Earlier
Previously, the context customization took advantage of an API exposed by
Spring Boot to customize the Tomcat Context.  It turns out that this happens
pretty late in the Context lifecycle, and specifically after Jasper scans for
TLDs.  This change updates the code to utilize a ContextLifecycleListener in
order ensure that it is called very early, at the expense of using a nicer
API.