Use ApplicationContextInitializers
Previously, this implementation contributed its modifications using an
AutoConfiguration hook.  It turns out that this is too late in the lifecycle
and the modifications needed an ApplicationContextInitializer to be early
enough.  Because of how early these contributions now happen, they are
responsible for doing their own class presence checking as well.